requirements.txt
astroid==2.3.3
Django==2.2.6
django-cors-headers==3.1.1
django-webpack-loader==0.6.0
djangorestframework==3.10.3
isort==4.3.21
mccabe==0.6.1
numpy==1.17.3
pylint==2.4.3
pylint-django==2.0.12
pylint-plugin-utils==0.6
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
